# billingwebapp/docs/openapi-ai-v1.yaml
openapi: 3.0.3
info:
  title: Clinic AI Integration API
  version: 1.0.0
  description: Secure server-to-server API for the separate Call Assistant. Direct database access is unsupported.
servers:
  - url: /api/v1/ai
security:
  - bearerAuth: []
paths:
...
  /appointments/slots:
    get:
      summary: Bookable slots after schedules, overrides, blocks and capacity
      x-required-scope: appointment:read
      parameters:
        - {name: doctor_id, in: query, required: true, schema: {type: integer, minimum: 1}}
        - {$ref: '#/components/parameters/LocationQuery'}
        - {$ref: '#/components/parameters/DateQuery'}
        - {name: start_time, in: query, schema: {type: string, pattern: '^\\d{2}:\\d{2}$'}}
        - {name: end_time, in: query, schema: {type: string, pattern: '^\\d{2}:\\d{2}$'}}
        - {name: period, in: query, schema: {type: string, enum: [morning, afternoon, evening]}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Conflict'}}
  /appointments:
    get:
      summary: Appointments owned by the verified patient
      x-required-scope: appointment:read
      parameters: [{$ref: '#/components/parameters/PatientSession'}, {name: appointment_no, in: query, schema: {type: string}}, {$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses: {'200': {$ref: '#/components/responses/Success'}}
    post:
      summary: Transactionally book a verified patient's slot
      x-required-scope: appointment:create
      parameters: [{$ref: '#/components/parameters/PatientSession'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AppointmentWrite'}}}}
      responses: {'201': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Conflict'}}
  /appointments/{appointment_no}/cancel:
    post:
      summary: Cancel an owned appointment without deleting history
      x-required-scope: appointment:update
      parameters: [{$ref: '#/components/parameters/AppointmentNo'}, {$ref: '#/components/parameters/PatientSession'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Conflict'}}
  /appointments/{appointment_no}/reschedule:
    post:
      summary: Transactionally move an owned appointment to a validated slot
      x-required-scope: appointment:update
      parameters: [{$ref: '#/components/parameters/AppointmentNo'}, {$ref: '#/components/parameters/PatientSession'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AppointmentWrite'}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '409': {$ref: '#/components/responses/Conflict'}}
  /appointments/{appointment_no}/late-arrival:
    post:
      summary: Record verified late-arrival state
      x-required-scope: appointment:update
      parameters: [{$ref: '#/components/parameters/AppointmentNo'}, {$ref: '#/components/parameters/PatientSession'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, properties: {status: {type: string, enum: [RUNNING_LATE, ARRIVED_LATE]}}}}}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /appointments/waitlist:
    post:
      summary: Add verified patient to optional waitlist
      x-required-scope: appointment:create
      parameters: [{$ref: '#/components/parameters/PatientSession'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'201': {$ref: '#/components/responses/Success'}}
...
components:
  securitySchemes:
    basicAuth: {type: http, scheme: basic}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque}
  parameters:
...
    LocationQuery: {name: location_id, in: query, required: true, schema: {type: integer, minimum: 1}}
    DateQuery: {name: date, in: query, required: true, schema: {type: string, format: date}}
    AppointmentNo: {name: appointment_no, in: path, required: true, schema: {type: string}}
...
    PatientSession: {name: X-Patient-Session, in: header, required: true, schema: {type: string}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 128}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
...
  schemas:
...
    AppointmentWrite:
      type: object
      required: [doctor_id, location_id, start_at]
      properties: {doctor_id: {type: integer}, location_id: {type: integer}, start_at: {type: string, format: date-time}, reason: {type: string, maxLength: 255}}
    Success:
      type: object
      required: [success, data, request_id]
      properties: {success: {type: boolean, enum: [true]}, data: {type: object}, request_id: {type: string}}
    Error:
      type: object
      required: [success, error, request_id]
      properties:
        success: {type: boolean, enum: [false]}
        error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}
        request_id: {type: string}
  responses:
    Success: {description: Standard success envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Success'}}}}
...
    Conflict: {description: State, capacity or idempotency conflict, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...